[kernel][arm64] comment and create separate versions of the way/set cache flushes
We will need more types of way/set caches flushes than just invalidate,
so move the logic of the routine into a macro and comment it to make it
easier to follow.

The logic comes directly from the ARM manual.

Tested: boot on qemu and real arm hardware, validate that flushes still
work.

Change-Id: I6bedc03c874c40829f11be220a6f66084030ea55