[kernel][mp][shutdown] rearrange the mp shutdown code to be a little clearer
The weak functions that were mostly already overloaded was not making
thigns easier to read and weren't particularly useful. Move the helper
thread function inside the caller as a lambda to be a little easier to
follow.

No functional change.

Tested: built and run on x86 and arm, verifying that mp shutdown still
works.

Change-Id: I3d68a327a8c96b5b3e79b15bccaa6ff6048fa837