[bt][gap-bredr] Drive link encryption with PairingState
Add a reference to an hci::Connection to PairingState. This provides the
link handle and link encryption handlers.

Implement PairingState kInitiatorWaitAuthComplete->kWaitEncryption and
kWaitLinkKey->kWaitEncryption transitions as a call to
hci::Connection::StartEncryption.

Implement OnEncryptionChange as a call to deliver a success status when
link encryption has been enabled.

BT-374

Test: in bt-host-unittests, GAP_PairingStateTest.*

Change-Id: Iaa59c4233bc99762d47bbb7f05175f4ee58a3a03