[camera][ISP] Moving frame processing from StatMgr to ArmIsp
Moves the thread which is signalled upon a new frame
entering the ISP from the Statmgr to the main ArmIspDevice
class. This is done so that the dma can be signalled when a
new frame arrives, and update the dma write address.

To test this CL, we must recieve signals from the ISP, meaning
the ISP and the sensor driver must be enabled.  A section of the
testing code is added to test the entire pipeline.

The StatMgr class only has the sensor instance in it now, meaning that
it can be removed.  For clarity, that will wait until a future CL.

Testability:  This CL adds to the ArmIspDeviceTester, which tests the
ISP on device.  For more information on how to run tests, see
//src/camera/TESTING.md

Change-Id: I65388db6e5d29030815252ac8e858bbf221b475d