[kernel][clocks] Align ticks with clock mono; bypass syscalls
Until now, user-mode vDSO images have always assumed that the tick
counter on x64 is sourced from the TSC, even if the TSC was not being
used as the source of time for clock monotonic.  This tended to be
because the TSC was readable from user-mode, but an alternate source
(like an HPET) would not be without significant effort to map the
HPET registers into user space as part of the vDSO.

Change this behavior so that ticks are always the same source of time
as clock monotonic.  This allows us to query clock monotonic by simply
transforming the ticks counter and without needing an explicit
syscall.  In order to achieve this, the reading of ticks on x64 may
need to be an actual syscall now, but only if a source of time _other_
than the TSC is chosen.

In practice, this should be quite rare.  The vast majority of x64
systems today have a proper invariant TSC, and will be able to read
either ticks or clock monotonic without any syscalls.  ARM systems use
the ARM generic timer, which is currently always accessible from user
mode.  Non-KVM QEMU-x64 systems are a notable example of a system
which does not expose an invariant TSC, and will still require a
syscall in order to read the ticks counter.

Next, remove the concept of the "soft tick counter".  Code which built
the vDSO image used to substitute in an implementation for ticks which
actually just mapped to clock monotonic, but only when the rate of the
tick counter was unknown at the time that the vDSO was being
constructed.  This is something which should _never_ happen.  The ARM
timers currently come up at INIT_LEVEL_PLATFORM_EARLY time, while x64
timers currently come up at INIT_LEVEL_VM + 3.  Both of these times
are significantly earlier than INIT_LEVEL_USER, which is the time when
the vDSO is constructed.  There is never a reason that we should not
know how fast our tick counter is running by the time we get to vDSO
construction.  Also remove the kernel command line flag which could be
used to force the kernel into soft tick counter mode.

Finally
++ Add kerenel a command line flags which can be used to _force_
   either ticks or clock monotonic to be queried using a syscall
   instead of register direct access.
++ Add kernel counters which can be used to track the number of
   syscalls made to zx_ticks_get, zx_clock_get_monotonic, and
   zx_clock_get.  In the case of zx_clock_get, subdivide the counter
   into the three types of clock currently supported.

Change-Id: Ibc02d52b4bae719cbfe7e447d1715f686a54e06a