[ISP] Moving first_frame to DmaMgr
The ArmIspDevice used to keep track of the first
time the dma would run, to have it signal that the previous
frame had been written.  Now the DmaMgr keeps track of this
variable, whichmay differ between the two dma's.

Testability:
This is tested by a new section of the ArmISPDeviceTester.
For more information about how to test the camera on-device,
see //src/camera/TESTING.md

Change-Id: Ie2a12c2c6f92c6e469dd6bbb8b8d88f24ce94f82