[graphics][compute]: Fix runtime crash on non-AMD GPUs.
This ensures that the VK_AMD_shader_info extension is only used
and called on AMD hardware only (if usage of the extension is enabled
through compiler macros, as is the case currently).

Before this, spinel_vk_svg would crash in the validation layers because
the program would try to call vkGetShaderInfoAMD even if the extension
was not enabled on my workstation (NVidia GPU).

+ Update the validation layer name in spinel_vk_context and
  spinel_vk_path_builder to let these programs run correctly
  (otherwise, they would fail early with an layer-not-present
  error).

  Note that while spinel_vk_context runs, spinel_vk_path_builder
  seems to loop endlessly with debug messages explaining that
  vkWaitForFences() is called with a |fenceCount| value of 0
  (which is an error according to the spec).

Testing: fx build src/graphics/lib/compute:all_tests &&
         out/default/host_x64/spinel_vk_svg

Change-Id: Ib9da5d8f53dd9f16d137f63211518e00e0e0d8d1