[graphics][compute][demo-sprint] Make spinel_vk_svg display its output.
Small CL to modify spinel_vk_svg to display its buffer in a window
(on the host) and on the framebuffer (on a device).

Works on the host, though the colors are wrong due to the format of the
swapchain images being different from what the buffer uses (easy to fix
in a future CL).

Does not work on the device though. Seems to be stuck during spinel
rendering on the Pixelbook. Probably requires a system update first
to update the Magma driver

Change-Id: Id7f3a9640ec0f8a1ccfed6e023cb35fd1bd95d02