[usb][sched] Apply deadline profiles to USB threads
Apply deadline profiles to the high priority XHCI IRQ completer, and to USB
audio threads.  Choose starting point numbers which reflect worst case scenarios
for each system.  These may need to be adjusted over time, but eventually should
be provided by a service with at least some knowledge of the platform, instead
of being synthesized by generic drivers.

Also, fix an error in the fifo depth reporting for audio outputs.  Right now, we
don't have direct DMA access to the USB host controller.  This means that the
audio driver needs to copy the data from the mixer ring buffer ahead of time,
and the fifo depth it report to the mixer needs to equal to the data in the
maximum number of packets in flight at any point in time.  Until now, it was
always being reported as two packets (eg; 2 mSec) even though it had been
leading the clock by 8.

Moving forward, fix the value that we report, and lower the max-in-flight number
to be just 3 packets.

Tested using instrumentation and trace logging on an Ava running with a simple
USB audio dongle and "loadgen" creating ugly amounts of load in the background.
Baring the strange PCIe command register read stuff (See bug 39839), IRQ latency
for USB maintained a steady ~5-6uSec, even when under load from 16 threads who
were behaving like jerks.

Change-Id: Iab2263c2f30bada855b469bd38460e820464bc98