[bt][emulator] Only make emulator visible once initialized
Ensures that the bt hci emulator device is not made visible until its
intialization is complete. This prevents race conditions between bind()
and unbind(), such as fxb/40794. The device model does not guarantee
that driver bind() will return before device unbind() is called. By
passing DEVICE_ADD_INVISIBLE to device_add(), we ensure that we can
continue safely initializing the device, and only device_make_visible()
once it is safe for unbind() to be called.

This fixes a flake in the set_active_host() test of
bt-integration-tests.

Bug: 40794

Test: $ fx run-test bt-integration-tests

Change-Id: I24d67e7afc9b06358d50cea633162673b367e66e