[audio] Implement FIDL interface for SimpleAudio users.
Implements the new FIDL audio driver interface for users of SimpleAudio.
The change is transparent for current drivers, i.e. their
usage of the data structures defined in audio.h is unchanged.
This does not prevent new driver to use FIDL directly for
more advanced funtionality than what is abstracted by SimpleAudio.
This change depends on the clients, i.e. audio_core and
Bluetooth, to add support for the new FIDL interface, i.e. it
won't be commited until those clients add the needed support.
This CL changes the node where drivers export the FIDL interface
such that it not possible to accidentally try to use the new
FIDL interface by a client using the manually serialized format
in audio.h.
This CL also removes use of audio-utils from SimpleAudio, other clients
of audio-utils like the audio uapp and hda still make use of it.

Bug: 35523

Test: runtests -t sa-unittest-test.
Test: audio uapp usage in following CL.
Test: E2E on Astro (once audio_core support fxb/41922 is added).

Change-Id: I1605ce06c766aee24ca70e6de585e8151118d481