[audio] Convert audio-utils to FIDL
On testability, after discussing with garratt@ I realized I should have
mentioned that I expect this library to not be used anymore by new code.
SimpleAudio used to use this library but it already moved away from
it (fxb/339030, it now calls the FIDL bindings directly). There are
still users though in particular the audio uapp utility
(system/uapp/audio) and HDA tests (system/dev/audio/intel-hda/test), and
hence this library is still needed. I'm testing it with the audio uapp
only with the expectation that once HDA moves away from this library,
we can remove it altogether, and hence investing in new tests for it
won't gives us much gain.

Test: audio uapp usage in following CL.
Change-Id: I450b3b7c6d2ec4aa33936bbfe3066f475ac90a0d