[audio] Add audio_core drivers FIDL interface
Adds to audio_core support for the FIDL audio driver API using
the HLCPP FIDL bindings. The manual serialization support is kept,
and it is chosen if defvs lists /dev/class/audio-input instead of
/dev/class/audio-input-2 (similarly for output).
Manual serialization support is renamed to AudioDriverLegacy and we
add a AudioDriverFidl class supporting FIDL. Note that we'll remove
AudioDriverLegacy once the transition is completed, so we keep some
duplication now to make the removal of AudioDriverLegacy easy.
Tests are moved from the legacy interface to the FIDL one via the
overloads to various interfaces (fake_audio_driver is
converted to the FIDL interface as well).

All tests below include the SimpleAudio transition to FIDL.

Test: fx run-test audio_core_unittests
Test: fx run-test audio_fidl_tests
Test: fx run-test audio_pipeline_test
Test: fx run-test audio_device_tests on Astro
Test: fx run-test audio_core_hardware_tests on Astro
Test: E2E audio input and output on Astro also without transition of SimpleAudio to FIDL
Change-Id: I3cbe745861caadc56f2eff27131b00bb96d4c90b