[audio] Add audio_core drivers FIDL interface
Adds to audio_core support for the FIDL audio driver API using
the HLCPP FIDL bindings. The manual serialization support is kept,
and it is chosen if defvs lists /dev/class/audio-input instead of
/dev/class/audio-input-2 (similarly for output).
The AudioDriverFidl class extends AudioDriver with the FIDL support.
Tests are moved from the legacy interface to the FIDL one via the
overloads to various interfaces (fake_audio_driver is
converted to the FIDL interface as well).

Updates to virtual audio are done in a differnt CL.

Test: fx run-test audio_core_unittests
Test: fx run-test audio_fidl_tests
Test: E2E audio input and output on Astro with and without transition of SimpleAudio to FIDL
Change-Id: I3cbe745861caadc56f2eff27131b00bb96d4c90b