[kernel][acpi_lite] add new lightweight library to parse necessary kernel ACPI tables
Add a simple library that replaces the simple part of finding the root
tables and decoding the simple ones that are used in the kernel.

Next change will wire it up to the acpi tables library which has test
code validating that the low level interface is working.

Testing: validate that it finds the ACPI table on a variety of hardware
including qemu and a few AMD and Intel machines.

Bug: <insert bug concerning replacing ACPICA with acpica lite>
Change-Id: I232dcde598e6c86df87557db476d2b863f865c37