[kernel][acpi] remove the syscall to transition to a particular S state
This code uses the last remnants of ACPICA which we're trying to remove
from the kernel. Will have to reimplement on top of the acpi_lite
library when the time comes to get it working again. The syspend syscall
hasn't worked in some time so just removing it so we can add it back
later is prudent.

Removes the last bit of ACPICA from the kernel.

Bug: <add bug referencing getting suspend back>
Bug: <insert bug concerning replacing ACPICA with acpica lite>
Change-Id: I2b3ef11a00e7abd568ea20cddb3d09efa5c78a7b