[audio] Add option for extending audio_core driver interface
Extend the AddDeviceByChannel FIDL API and various internal audio_core
interfaces with overloads to differentiate instances of the
driver interface supporting the legacy manual serializations versus
the FIDL based interface. Subsequent CLs add the actual support for
the FIDL interface and all unit tests are kept with the legacy interface.

We also move SelectBestFormat as a driver method, this avoids users
spefifying format_ranges() that returns a type specify to the legacy
implementation.

The devfs detection is also extended to support the
/dev/class/audio-input-2 and /dev/class/audio-output-2 nodes needed for
the FIDL interface.

In order to overload AddDeviceByChannel FIDL definition we need to
import the audio driver FIDL definitions, due to this we need to flag
the audio driver interface as sdk_category = partner same as
audio_core's FIDL.

Test: fx run-test audio_core_unittests
Test: fx run-test audio_device_service_tests
Test: E2E audio input and output on Astro
Change-Id: Ibbb2b2c0dd8715dde36f4d689f09ebb32d9ddbf6