[shutdown-shim] reduce statecontrol.Admin timeout
The shutdown-shim's timeout for power_manager is always waited for on
bringup, as power_manager will never finish loading on bringup. To allow
developers to reboot in < 20 seconds on bringup, reduce this timeout to
2 seconds. In all typical shutdown scenarios power_manager will already
be running, so it should take far less than 2 seconds for the connection
to be established with power_manager.

Change-Id: Ic01e521dde162e84827da3010b462a4752474cc8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400815
Reviewed-by: Patrick Shickel <pshickel@google.com>
Testability-Review: Derek Gonyeo <dgonyeo@google.com>
Commit-Queue: Derek Gonyeo <dgonyeo@google.com>