[tee][optee] Lower log level of OpteeClient logs
The OpteeClient acts as a proxy between a Trusted App and a Client App.
If the TEE OS, Trusted App or Client App provide malformed or
unsupported messages, this does not indicate a bug in the driver. As
such, the driver should rarely be using the ERROR level for these logs.
This CL updates the log statements in OpteeClient to mostly use the
DEBUG level.

Change-Id: Ic4c48f2e3e1cc0df4f09b8c84eef88f5dd3676fa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401976
Reviewed-by: Austin Foxley <afoxley@google.com>
Reviewed-by: Chris Tam <godtamit@google.com>
Testability-Review: Austin Foxley <afoxley@google.com>
Commit-Queue: RJ Ascani <rjascani@google.com>