[fx][test] fix hash pinning update on build
fxtest was fetching the merkle root hash in a wrong location, before
the build was attempted. As a consequence, it was using the hash of the
package before the build, hence executing stale tests.

Implementation details of this fix: executionhandle is the objec that
contains the exact Fuchsia URL of the component test. However, it was
being created at testdefinition construction time, which is way before
the 'fx build' has a chance to run. In order to fix the issue, this CL:
- moves testType from ExecutionHandle to TestDefinition
- adds a "hash" public non-final attribute to TestDefinition
- moves ExecutionHandle from being an attribute set at construction time
in TestDefinition to being a public method that calculates it at
execution time, and makes that method use the "hash" that is set at the
moment of execution (via the new decoratedPackageUrl getter)

As a side effect, 'fxPath' is not needed anymore in TestDefinition.
Now, instead of having the full path of fx in CommandTokens, only the
string 'fx' is used, and TestBundle._runTest does two new transformations:
replace 'fx' with FxLocator.fxPath if 'fx' is the first element of the
CommandTokens, and query the PackageRepository for the
testBundle.testDefinition packageName and set it in testDefinition.hash.

Fixed: 54881
Tests:
  - added a new integration test in fx self-test to verify this behavior
  - manually tested with a component test
  - 'fx test fxtest'
  - 'fx self-test subcommands/fx_test_test'

Change-Id: I44ae28f54c8f162458a2462413fcc7b19e9b9555
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401794
Commit-Queue: Renato Mangini Dias <mangini@google.com>
Testability-Review: Renato Mangini Dias <mangini@google.com>
Testability-Review: Craig Labenz <craiglabenz@google.com>
Reviewed-by: Craig Labenz <craiglabenz@google.com>