[storage] Move minfs tests to new fs_test framework.
Also fixed file_system -> filesystem and a header path that was wrong.

Tweaked the test slightly so that it runs a little faster which will
allow us to always run the FullOperations test (which was previously
marked as large so didn't run in CQ), and the test actually got broken
recently (when the minimum inode count was reduced) and will no longer
run on the non-FVM version of the test. That probably doesn't matter; we
have enough coverage from the FVM version of the test.

Bug: 53179
Change-Id: Icd1274ee0a6d07f8d9be2c172e4eec950a08fdb1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/401893
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Testability-Review: Chris Suter <csuter@google.com>